fix(api): DEEP_RESILIENCE V6.0 MEDIO — PSP state machine + split range
- #DR263 PspCobrancaService: whitelist de transicoes de status (PENDENTE->CONFIRMADA/VENCIDA/CANCELADA, terminais ESTORNADA/CANCELADA absorventes) — webhook fora-de-ordem nao corrompe mais o status final.
- #DR264 AsaasGateway.criarCobranca: valida splitNavieraPct in [0,100] antes do POST (mensagem opaca do Asaas vira IllegalArgumentException com contexto).
- OpEncomendaWriteService.entregar: sigatura Long->Integer para casar com TenantUtils.getOperadorId (compile error pre-existente).
- OnboardingService.registrarEmpresa: cnpjFinal final var para satisfazer captura em lambda.

+++ naviera-api/src/main/java/com/naviera/api/service/OpEncomendaWriteService.java
@@ -122,7 +122,7 @@ public class OpEncomendaWriteService {
     }
 
     @Transactional
-    public Map<String, Object> entregar(Integer empresaId, Long operadorId, Long id, Map<String, Object> dados) {
+    public Map<String, Object> entregar(Integer empresaId, Integer operadorId, Long id, Map<String, Object> dados) {
         // #DS5-020: validar identificacao do recebedor antes de marcar como entregue.
         String docRec = trimToNull(dados.get("doc_recebedor"));
         String nomeRec = trimToNull(dados.get("nome_recebedor"));
